+++ Packages/com.github.kurotu.vrc-quest-tools/Editor/Inspector/AvatarConverterSettingsEditor.cs
@@ -102,23 +102,12 @@ namespace KRT.VRCQuestTools.Inspector
                     {
                         using (var horizontal = new EditorGUILayout.HorizontalScope())
                         {
-                            EditorGUILayout.HelpBox(i18n.AlertForMissingNetIds, MessageType.Warning);
-                            if (GUILayout.Button(i18n.AssignButtonLabel, GUILayout.Height(38), GUILayout.Width(60)))
-                            {
-                                OnClickAssignNetIdsButton(descriptor);
-                            }
-                        }
-
-#if VQT_HAS_NDMF
-                        using (var horizontal = new EditorGUILayout.HorizontalScope())
-                        {
                             EditorGUILayout.HelpBox(i18n.InfoForNetworkIdAssigner, MessageType.Info);
                             if (GUILayout.Button(i18n.AttachButtonLabel, GUILayout.Height(38), GUILayout.Width(60)))
                             {
                                 OnClickAttachNetworkIDAssignerButton(descriptor);
                             }
                         }
-#endif
                     }
 
                     var pbs = descriptor.GetComponentsInChildren(VRCSDKUtility.PhysBoneType, true);
@@ -276,6 +265,7 @@ namespace KRT.VRCQuestTools.Inspector
                 {
                     EditorGUILayout.PropertyField(so.FindProperty("animatorOverrideControllers"), new GUIContent(i18n.AnimationOverrideLabel, i18n.AnimationOverrideTooltip));
                     EditorGUILayout.PropertyField(so.FindProperty("removeVertexColor"), new GUIContent(i18n.RemoveVertexColorLabel, i18n.RemoveVertexColorTooltip));
+                    EditorGUILayout.PropertyField(so.FindProperty("ndmfPhase"), new GUIContent(i18n.NdmfPhaseLabel, i18n.NdmfPhaseTooltip));
                 }
 
                 Views.EditorGUIUtility.HorizontalDivider(2);
@@ -448,6 +438,7 @@ namespace KRT.VRCQuestTools.Inspector
         {
             descriptor.gameObject.AddComponent<NetworkIDAssigner>();
             PrefabUtility.RecordPrefabInstancePropertyModifications(descriptor.gameObject);
+            EditorUtility.DisplayDialog(VRCQuestTools.Name, i18n.NetworkIdAssignerAttached, "OK");
         }
 
         private void OnClickRegenerateTexturesButton(VRC_AvatarDescriptor avatar, IMaterialConvertSettings convertSetting)