+++ Packages/com.github.kurotu.vrc-quest-tools/Runtime/Models/AvatarConverterNdmfPhase.cs
@@ -1,0 +1,14 @@
+namespace KRT.VRCQuestTools.Models
+{
+    /// <summary>
+    /// NDMF phase for AvatarConverterSettings.
+    /// </summary>
+    public enum AvatarConverterNdmfPhase
+    {
+        /// <summary>Transforming phase.</summary>
+        Transforming,
+
+        /// <summary>Optimizing phase.</summary>
+        Optimizing,
+    }
+}